# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WeePlan API",
    "version": "1.0.0",
    "description": "API documentation for WeePlan Backend"
  },
  "servers": [
    {
      "url": "http://localhost:5013",
      "description": "Local server"
    },
    {
      "url": "http://3.226.133.13",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register User or Vendor",
        "description": "Register a new user or vendor account.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User registration details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "phone": {
                  "type": "string",
                  "example": "01700000000"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                },
                "acceptedTerms": {
                  "type": "boolean",
                  "example": true
                },
                "role": {
                  "type": "string",
                  "example": "user"
                },
                "fcmToken": {
                  "type": "string",
                  "example": "optional_token"
                }
              },
              "required": [
                "firstName",
                "lastName",
                "email",
                "phone",
                "password",
                "acceptedTerms",
                "role"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "body": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/resendOtp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Resend OTP",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "identifier": {
                  "type": "string",
                  "example": "user@example.com"
                }
              },
              "required": [
                "identifier"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User Login",
        "description": "Login using email or phone identifier along with password.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User login credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "identifier": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                },
                "fcmToken": {
                  "type": "string",
                  "example": "optional_token"
                }
              },
              "required": [
                "identifier",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/auth/admin/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Admin Login",
        "description": "Login for admin and superAdmin.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Admin login credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "identifier": {
                  "type": "string",
                  "example": "admin@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                },
                "fcmToken": {
                  "type": "string",
                  "example": "optional_token"
                }
              },
              "required": [
                "identifier",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/auth/changePassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Change Password",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "oldPassword": {
                  "type": "string",
                  "example": "old_password_123"
                },
                "newPassword": {
                  "type": "string",
                  "example": "new_password_123"
                }
              },
              "required": [
                "oldPassword",
                "newPassword"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/auth/refresh-token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh Token",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "type": "string",
                  "example": "your_refresh_token_here"
                }
              },
              "required": [
                "refreshToken"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/auth/forgotPass": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot Password",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "phone": {
                  "type": "string",
                  "example": "01700000000"
                }
              },
              "required": [
                "phone"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/resetPass": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset Password",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "phone": {
                  "type": "string",
                  "example": "01700000000"
                },
                "otp": {
                  "type": "string",
                  "example": "123456"
                },
                "newPassword": {
                  "type": "string",
                  "example": "new_password_123"
                }
              },
              "required": [
                "phone",
                "otp",
                "newPassword"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/auth/regOtpVerify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify Registration OTP",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "identifier": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "otp": {
                  "type": "string",
                  "example": "123456"
                }
              },
              "required": [
                "identifier",
                "otp"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "example": "any"
                  },
                  "otp": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin Login",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "identifier": {
                  "type": "string",
                  "example": "admin@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              },
              "required": [
                "identifier",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}